refactor: split introduction in more components
