+++ packages/api-reference/src/components/ApiReferenceLayout.vue
@@ -42,7 +42,7 @@ import { hasLazyLoaded } from '@/components/Lazy/lazyBus'
 import MobileHeader from '@/components/MobileHeader.vue'
 import { ApiClientModal } from '@/features/api-client-modal'
 import { useDocumentSource } from '@/features/document-source'
-import { OPENAPI_VERSION_SYMBOL } from '@/features/download-link'
+import { OPENAPI_VERSION_SYMBOL } from '@/features/download-links'
 import { SearchButton } from '@/features/Search'
 import { Sidebar, useSidebar } from '@/features/sidebar'
 import { CONFIGURATION_SYMBOL } from '@/hooks/useConfig'
