+++ packages/api-reference/src/components/Content/Introduction/IntroductionSection.test.ts
@@ -1,4 +1,4 @@
-import { DownloadLink } from '@/features/download-link'
+import { DownloadLink } from '@/features/download-links'
 import { useSidebar } from '@/features/sidebar/hooks/useSidebar'
 import type { OpenApiDocument } from '@scalar/workspace-store/schemas/v3.1/strict/openapi-document'
 import { mount } from '@vue/test-utils'
