+++ packages/api-reference/src/components/Content/Introduction/IntroductionSection.vue
@@ -1,10 +1,11 @@
 <script setup lang="ts">
 import type { ApiReferenceConfiguration } from '@scalar/types/api-reference'
 import type { OpenApiDocument } from '@scalar/workspace-store/schemas/v3.1/strict/openapi-document'
-import { computed, inject, onMounted, type Ref } from 'vue'
+import { onMounted } from 'vue'
 
-import { Badge } from '@/components/Badge'
-import { LinkList } from '@/components/LinkList'
+import ScalarInfoLinks from '@/components/Content/Introduction/ScalarInfoLinks.vue'
+import ScalarInfoVersion from '@/components/Content/Introduction/ScalarInfoVersion.vue'
+import ScalarOpenApiVersion from '@/components/Content/Introduction/ScalarOpenApiVersion.vue'
 import {
   Section,
   SectionColumn,
@@ -14,14 +15,12 @@ import {
   SectionHeader,
   SectionHeaderTag,
 } from '@/components/Section'
-import { DownloadLink, OPENAPI_VERSION_SYMBOL } from '@/features/download-link'
-import { ExternalDocs } from '@/features/external-docs'
-import { Contact, License, TermsOfService } from '@/features/info-object'
+import { DownloadLinks } from '@/features/download-links'
 import { SpecificationExtension } from '@/features/specification-extension'
 import { DEFAULT_INTRODUCTION_SLUG } from '@/features/traverse-schema'
 import { useNavState } from '@/hooks/useNavState'
 
-import Description from './Description.vue'
+import ScalarInfoDescription from './ScalarInfoDescription.vue'
 
 const { document, config } = defineProps<{
   document: OpenApiDocument
@@ -30,27 +29,10 @@ const { document, config } = defineProps<{
 
 const { getHeadingId } = useNavState()
 
-/**
- * Get the OpenAPI/Swagger specification version from the API definition.
- */
-const oasVersion = inject<Ref<string | undefined>>(OPENAPI_VERSION_SYMBOL)
-
-/** Format the version number to be displayed in the badge */
-const version = computed(() => {
-  // Prefix the version with “v” if the first character is a number, don't prefix if it's not.
-  // Don't output anything when version is not a string.
-  return typeof document.info?.version === 'string'
-    ? document.info.version.toString().match(/^\d/)
-      ? `v${document.info.version}`
-      : document.info.version
-    : typeof document.info?.version === 'number'
-      ? `v${document.info.version}`
-      : undefined
-})
-
 /** Trigger the onLoaded event when the component is mounted */
 onMounted(() => config?.onLoaded?.())
 </script>
+
 <template>
   <SectionContainer>
     <!-- If the #after slot is used, we need to add a gap to the section. -->
@@ -69,8 +51,8 @@ onMounted(() => config?.onLoaded?.())
           (!document?.info?.description && !document?.info?.title)
         ">
         <div class="flex gap-1.5">
-          <Badge v-if="version">{{ version }}</Badge>
-          <Badge v-if="oasVersion">OAS {{ oasVersion }}</Badge>
+          <ScalarInfoVersion :version="document.info?.version" />
+          <ScalarOpenApiVersion />
         </div>
         <SectionHeader
           :loading="!document.info?.title"
@@ -79,26 +61,13 @@ onMounted(() => config?.onLoaded?.())
             {{ document.info?.title }}
           </SectionHeaderTag>
           <template #links>
-            <LinkList>
-              <ExternalDocs :value="document.externalDocs" />
-              <Contact
-                v-if="document.info?.contact"
-                :value="document.info?.contact" />
-              <License
-                v-if="document.info?.license"
-                :value="document.info?.license" />
-              <TermsOfService
-                v-if="document.info?.termsOfService"
-                :value="document.info?.termsOfService" />
-            </LinkList>
+            <ScalarInfoLinks :document="document" />
           </template>
         </SectionHeader>
         <SectionColumns>
           <SectionColumn>
-            <div class="links">
-              <DownloadLink :title="document.info?.title" />
-            </div>
-            <Description :value="document.info?.description" />
+            <DownloadLinks :title="document.info?.title" />
+            <ScalarInfoDescription :description="document.info?.description" />
           </SectionColumn>
           <SectionColumn v-if="$slots.aside">
             <div class="sticky-cards">
