+++ packages/api-reference/src/features/download-links/index.ts
@@ -1,2 +1,2 @@
-export { default as DownloadLink } from './DownloadLink.vue'
+export { default as DownloadLinks } from './DownloadLinks.vue'
 export { OPENAPI_VERSION_SYMBOL } from './openapi-version-symbol'
