+++ packages/api-reference/src/features/download-links/openapi-version-symbol.ts
